+++ extensions/Worker.Extensions.CosmosDB/src/Config/CosmosDBBindingOptions.cs
@@ -9,6 +9,12 @@ using Microsoft.Azure.Functions.Worker.Extensions.CosmosDB;
 
 namespace Microsoft.Azure.Functions.Worker
 {
+     /// <summary>
+    /// Internal options for configuring the CosmosDB binding.
+    /// This class is used internally by the Azure Functions runtime to manage the CosmosDB connection and clients.
+    /// It is not intended to be used directly in user code.
+    /// Any public configuration options should be set on the <see cref="CosmosDBExtensionOptions"/> class, which is publicly accessible.
+    /// </summary>
     internal class CosmosDBBindingOptions
     {
         public string? ConnectionName  { get; set; }
@@ -19,7 +25,7 @@ namespace Microsoft.Azure.Functions.Worker
 
         public TokenCredential? Credential { get; set; }
 
-        public CosmosSerializer? Serializer { get; set; }
+        public CosmosDBExtensionOptions? CosmosExtensionOptions { get; set; }
 
         internal string BuildCacheKey(string connection, string region) => $"{connection}|{region}";
 
@@ -32,24 +38,23 @@ namespace Microsoft.Azure.Functions.Worker
                 throw new ArgumentNullException(nameof(ConnectionName));
             }
 
+            if (CosmosExtensionOptions is null)
+            {
+                CosmosExtensionOptions = new CosmosDBExtensionOptions();
+            }
+
             string cacheKey = BuildCacheKey(ConnectionName!, preferredLocations);
 
-            CosmosClientOptions cosmosClientOptions = new ()
+            // Do not override if preferred locations is configured via CosmosClientOptions
+            if (!string.IsNullOrEmpty(preferredLocations) && CosmosExtensionOptions.ClientOptions.ApplicationPreferredRegions is null)
             {
-                ConnectionMode = ConnectionMode.Gateway
-            };
-
-            if (!string.IsNullOrEmpty(preferredLocations))
-            {
+                // Clone to avoid modifying the original options
+                CosmosClientOptions cosmosClientOptions = CosmosExtensionOptions.ClientOptions.MemberwiseClone<CosmosClientOptions>();
                 cosmosClientOptions.ApplicationPreferredRegions = Utilities.ParsePreferredLocations(preferredLocations);
+                return ClientCache.GetOrAdd(cacheKey, (c) => CreateService(cosmosClientOptions));
             }
 
-            if (Serializer is not null)
-            {
-                cosmosClientOptions.Serializer = Serializer;
-            }
-
-            return ClientCache.GetOrAdd(cacheKey, (c) => CreateService(cosmosClientOptions));
+            return ClientCache.GetOrAdd(cacheKey, (c) => CreateService(CosmosExtensionOptions.ClientOptions));
         }
 
         private CosmosClient CreateService(CosmosClientOptions cosmosClientOptions)