+++ extensions/Worker.Extensions.CosmosDB/release_notes.md
@@ -4,8 +4,19 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB 4.8.1
+### Microsoft.Azure.Functions.Worker.Extensions.CosmosDB 4.9.0
 
-- Updating `Microsoft.Azure.WebJobs.Extensions.CosmosDB` reference to 4.6.1
-- Updating `Microsoft.Extensions.Azure` reference to 1.7.3
-- Updating `Microsoft.Azure.Cosmos` reference to 3.39.1
+- Implement `CosmosDBExtensionOptions` to allow configuration of the CosmosDB service client via `CosmosClientOptions` (#2483)
+
+#### Example Usage
+
+```csharp
+.ConfigureFunctionsWorkerDefaults((builder) =>
+{
+    builder.ConfigureCosmosDBExtensionOptions((options) =>
+    {
+        options.ClientOptions.ConnectionMode = ConnectionMode.Direct;
+        options.ClientOptions.ApplicationName = "MyApp";
+    });
+})
+```